ScoreHandler:
- can now take in negative scores (no longer size_t but int64_t)
- Fix a bug where it would overwrite itself

Minesweeper:
- Menu bar
- Save score when game is quit
- Limited chrono time
- Chrono and Flags/Bombs on menu bar
- Only flag tiles that aren't revealed
- Get rid of state variables for an enum GameState
- Victory condition

Closes #3

+++ include/games/ScoreHandler.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <cstdint>
 #include <fstream>
 #include <string>
 #include <vector>
@@ -12,12 +13,12 @@ namespace arcade {
 
             ~ScoreHandler();
 
-            void addScore(std::size_t plusScore);
-            std::size_t getScore() const;
+            void addScore(std::int64_t plusScore);
+            std::int64_t getScore() const;
             void resetScore();
 
             // Pair of player names and scores
-            std::vector<std::pair<const std::string, const std::size_t>> loadScores();
+            std::vector<std::pair<const std::string, const std::int64_t>> loadScores();
 
             void saveScore(const std::string playerName);
 
@@ -29,7 +30,7 @@ namespace arcade {
             std::ifstream _fileInput;
 
             const std::string _gameName;
-            std::size_t _score;
+            std::int64_t _score;
 
             bool _isSaved;
